Profit Curve - Minima and Maxima
This commit entails analyzing a time series plot to find the local maxima and minima in order to get the magnitudes of price drops of daily equity price. This price-equity curve contains several continuous price drops: sequences between a local price maximum and a subsequent local price minimum in which the price is either falling or staying fixed. The magnitude of a price drop is the difference in the price at the local maximum minus the price at the local minimum.